requirements.txt
affine==2.4.0
cartopy==0.25.0
earthaccess==0.16.0
faiss==1.5.3
harmony_py==1.3.3
langchain==1.2.10
langchain_core==1.2.16
langchain_huggingface==1.2.0
langchain_ollama==1.0.1
matplotlib==3.10.8
netCDF4==1.7.2
netCDF4==1.7.4
numpy==2.4.2
parsedatetime==2.6
python-dotenv==1.2.2
rasterio==1.4.3
Requests==2.32.5
sentence_transformers==5.2.0
Shapely==2.1.2
spacy==3.8.11
xarray==2025.6.1
zarr==3.1.5
